Rename ArticuloManufacturadoForm -> ManufacturadoForm
