+++ src/components/domain/producto/ManufacturadoForm.tsx
